# src/twitter-swagger.yaml
openapi: 3.0.3
info:
  title: twitter-clone APIs
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
    Some useful links:
    - [how to use Swagger](https://youtu.be/vhFK6hXV-Qo)
    - [The source API definition for twitter-clone](https://github.com/darknesshaotran/Twitter-Clone)
  contact:
    email: darknessclmm@gmail.com
  version: 1.0.11
servers:
  - url: http://localhost:3000
...
paths:
...
  /users/refresh_token:
    post:
      tags:
        - users
      summary: Refresh token
      description: refresh token every time access token expire
      operationId: refresh_token
      requestBody:
        description: refreshToken's request
        content:
          application/json:
            schema: 
              type: object
              properties:
                Refresh_token:
                  type: string
                  format: JWT
                  example:  eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NGI1NjFjNzJhOTE0Yjc4NWE1NjkzZDYiLCJ2ZXJpZnkiOjEsInR5cGUiOjAsImlhdCI6MTY5MzY0MzQ4NiwiZXhwIjoxNjkzNjQ1Mjg2fQ.SxPMJ8a6zW38FH5CCmCAg9yu38EjnFrFOqh0N8RKDvg       
        required: true
      responses:
        default:
          description: refresh token is successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: refresh token is successfully
                  result:
                    $ref: '#/components/schemas/SuccessAuthentication'    
...
components:
  schemas:
...
    SuccessAuthentication:
      type: object
      properties:
        AccessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NGMwODU1NTBkYjVhM2VlZTRhY2ZlNjIiLCJ2ZXJpZnkiOjEsInR5cGUiOjAsImlhdCI6MTY5MzYyODg1NywiZXhwIjoxNjkzNjMwNjU3fQ.ggS4Via1jNEVptrwy9ckRtguSJJaW6rEUachOuxLgfY
        Refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2NGMwODU1NTBkYjVhM2VlZTRhY2ZlNjIiLCJ2ZXJpZnkiOjEsInR5cGUiOjEsImlhdCI6MTY5MzYyODg1NywiZXhwIjoxNzAyMjY4ODU3fQ.q_kFBjtisjhMHcOrwnd06tWtVla0xmOj8leVcPme2qQ
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT